% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppsilhobj.R
\name{ppsilhobj}
\alias{ppsilhobj}
\alias{plot.ppsilhobj}
\title{Calculate Silhouette Width for Soft Clustering Algorithms using Cluster Membership Probability}
\usage{
ppsilhobj(
  pp,
  pptype = c("pp", "nlpp", "pd"),
  silhmethod = c("silh", "pac", "cer"),
  average = c("crisp", "fuzzy"),
  a = 2,
  clust_function = NULL,
  ...
)

\method{plot}{ppsilhobj}(ppsilhobj, label = FALSE, print.summary = TRUE, ...)
}
\arguments{
\item{pp}{A matrix representing cluster membership probabilities.}

\item{pptype}{Character string specifying the type of membership matrix:
\describe{
\item{"pp"}{Posterior Probability}
\item{"nlpp"}{Negative Log of Posterior Probability}
\item{"pd"}{Probability Distribution}
}}

\item{silhmethod}{Character string indicating the silhouette calculation method:
\describe{
\item{"silh"}{Standard silhouette calculation}
\item{"pac"}{Poserior Probability of adjucent cluster}
\item{"cer"}{Certainty}
}}

\item{average}{Character string specifying the type of average silhouette width calculation:
\describe{
\item{"crisp"}{Simple average of silhouette widths}
\item{"fuzzy"}{Weighted average based on membership probabilities}
}}

\item{a}{Numeric value controlling the weight calculation based on membership difference.}

\item{clust_function}{Optional S3 function name (as a string) to produce the proximity measure matrix.}

\item{...}{Additional arguments passed to \code{ggpubr::ggpar}.}

\item{ppsilhobj}{An object of class \code{ppsilhobj}, created by \code{ppsilhobj}}

\item{label}{A logical value indicating whether to label the x-axis with observation names. Defaults to FALSE.}

\item{print.summary}{A logical value indicating whether to print a summary of average silhouette widths for each cluster. Defaults to TRUE.}
}
\value{
A list with the following components:
\describe{
\item{widths}{A data frame containing cluster assignments, nearest neighbor, weight, and silhouette widths.}
\item{clus.avg.widths}{A named vector of average silhouette widths for each cluster.}
\item{avg.width}{The overall average silhouette width across all observations.}
}

A ggplot2 object representing the silhouette plot.
}
\description{
This function calculates the silhouette width for each observation based on the clustering results.
The silhouette width measures how similar an observation is to its own cluster compared to other clusters.
}
\examples{
library(cluster)
out <- fanny(iris[,-5], 3)

silhout1 <- ppsilhobj(out$membership,average = "fuzzy",pptype = "pd")
plot(silhout1)

silhout2 <- ppsilhobj(out$membership,average = "crisp",pptype = "pd")
plot(silhout2)
}
