% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silh.R
\name{silh}
\alias{silh}
\title{Calculate Silhouette Width for Clustering and Block clustering}
\usage{
silh(
  pm,
  pmtype = c("sim", "dissim"),
  silhmethod = c("silh", "pac"),
  clust_function = NULL,
  ...
)
}
\arguments{
\item{pm}{A matrix or list of matrices where each row represents an observation and each column represents a proximity measure
to different clusters. Typically, this matrix would be a membership matrix or dissimilarity matrix from clustering results or their variable names string in vector.}

\item{pmtype}{Character string indicating whether \code{pm} represents similarity ("sim") or dissimilarity ("dissim").}

\item{silhmethod}{Character string indicating the silhouette calculation method: "pac" or "silh".}

\item{clust_function}{Optional S3 function name (as a string) that produces the proximity measure matrix.}

\item{...}{Additional arguments passed to \code{clust_function}, if used.}
}
\value{
A list or two lists with the following components for row and column proximity measures:
\itemize{
\item \code{widths}: A data frame containing the cluster assignments, nearest neighbor, and silhouette widths.
\item \code{clus.avg.widths}: A named vector of average silhouette widths for each cluster.
\item \code{avg.width}: The overall average silhouette width across all observations.
}

\code{block.shil.matrix} : A matrix containing silhout index of each data point from block clustering (if inputs are from block custering/ Bi-clustering/Coclustering).
}
\description{
This function calculates the silhouette width for each observation based on the clustering or block clustering results.
The silhouette width is used to measure how similar an observation is to its own cluster compared
to other clusters.
}
