% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silhouette-package.R
\docType{package}
\name{Silhouette-package}
\alias{Silhouette-package}
\title{Silhouette: Proximity Measure Based Diagnostics for Standard, Soft, and Multi-Way Clustering}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Quantifies clustering quality by measuring both cohesion within clusters and separation between clusters. Implements advanced silhouette width computations for diverse clustering structures, including: simplified silhouette (Van der Laan et al., 2003) \doi{10.1080/0094965031000136012}, Probability of Alternative Cluster normalization methods (Raymaekers & Rousseeuw, 2022) \doi{10.1080/10618600.2022.2050249}, fuzzy clustering and silhouette diagnostics using membership probabilities (Campello & Hruschka, 2006; Menardi, 2011; Bhat & Kiruthika, 2024) \doi{10.1016/j.fss.2006.07.006}, \doi{10.1007/s11222-010-9169-0}, \doi{10.1080/23737484.2024.2408534}, and multi-way clustering extensions such as block and tensor clustering (Schepers et al., 2008; Bhat & Kiruthika, 2025) \doi{10.1007/s00357-008-9005-9}, \doi{10.21203/rs.3.rs-6973596/v1}. Provides tools for computation and visualization (Rousseeuw, 1987) \doi{10.1016/0377-0427(87)90125-7} to support robust and reproducible cluster diagnostics across standard, soft, and multi-way clustering settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kskbhat.github.io/Silhouette/}
  \item Report bugs at \url{https://github.com/kskbhat/Silhouette/issues}
}

}
\author{
\strong{Maintainer}: Shrikrishna Bhat K \email{skbhat.in@gmail.com} (\href{https://orcid.org/0009-0000-6180-5783}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kiruthika C (\href{https://orcid.org/0000-0001-9655-702X}{ORCID})
}

}
\keyword{internal}
