% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerSilhouette.R
\name{cerSilhouette}
\alias{cerSilhouette}
\title{Certainty Silhouette Width (Cer) for Soft Clustering}
\usage{
cerSilhouette(
  prob_matrix,
  average = c("crisp", "fuzzy", "median"),
  a = 2,
  sort = FALSE,
  print.summary = FALSE,
  clust_fun = NULL,
  ...
)
}
\arguments{
\item{prob_matrix}{A numeric matrix of posterior probabilities where rows represent observations and columns represent clusters. Must sum to 1 by row. If \code{clust_fun} is provided, \code{prob_matrix} must be a string giving the name of the matrix component (e.g., "u").}

\item{average}{Character string specifying the method for computing the average silhouette width.
Options are:
\itemize{
\item \code{"crisp"} – unweighted (simple) average.
\item \code{"fuzzy"} – weighted average based on membership probability differences.
\item \code{"median"} – median silhouette width across observations.
}
Defaults to \code{"crisp"}.}

\item{a}{Numeric value controlling the fuzzifier or weight scaling in fuzzy silhouette averaging. Higher values increase the emphasis on strong membership differences. Must be positive. Defaults to \code{2}.}

\item{sort}{Logical; if \code{TRUE}, sorts the output \code{widths} data frame by cluster and descending silhouette width. Defaults to \code{FALSE}.}

\item{print.summary}{Logical; if \code{TRUE}, prints a summary table of average silhouette widths and sizes for each cluster. Defaults to \code{FALSE}.}

\item{clust_fun}{Optional S3 or S4 function object or function as character string specifying a clustering function that produces the proximity matrix. For example, \code{\link[ppclust]{fcm}} or \code{"fcm"}. If provided, \code{prob_matrix} must be a string giving the name of the matrix component (e.g., "u"). Defaults to \code{NULL}.}

\item{...}{Additional arguments passed to \code{clust_fun}, such as \code{x, centers} for \code{\link[ppclust]{fcm}}.}
}
\value{
A data frame of class \code{"Silhouette"} containing cluster assignments, nearest neighbor clusters, silhouette widths for each observation, and weights (for fuzzy clustering). The object includes the following attributes:
\describe{
\item{proximity_type}{The proximity type used i.e., \code{"similarity"}.}
\item{method}{The silhouette calculation method used i.e., \code{"certainty"}.}
\item{\code{"average"}}{Character — the averaging method: \code{"crisp"}, \code{"fuzzy"}, or \code{"median"}.}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes silhouette widths using maximum of posterior probabilities as Silhouette.
}
\details{
Let the posterior probability matrix as
\deqn{\Gamma = [\gamma_{ik}]_{n \times K},}
The \strong{certainty silhouette width} for observation \eqn{i} is:
\deqn{
\mathrm{Cer}_i = \max_{k=1,\dots,K} \gamma_{ik}
}
}
\references{
Bhat Kapu, S., & Kiruthika. (2024). Some density-based silhouette diagnostics for soft clustering algorithms. Communications in Statistics: Case Studies, Data Analysis and Applications, 10(3-4), 221-238. \doi{10.1080/23737484.2024.2408534}
}
\seealso{
\code{\link{Silhouette}},\code{\link{softSilhouette}},\code{\link{dbSilhouette}},\code{\link{plotSilhouette}},
}
