% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{Calculate Silhouette Width for Clustering}
\usage{
silhouette(
  prox_matrix,
  proximity_type = c("similarity", "dissimilarity"),
  method = c("medoid", "pac"),
  prob_matrix = NULL,
  a = 2,
  clust_fun = NULL,
  ...
)
}
\arguments{
\item{prox_matrix}{A numeric matrix where rows represent observations and columns represent proximity measures (e.g., distances or similarities) to clusters. Typically, this is a membership or dissimilarity matrix from clustering results. If \code{clust_fun} is provided, \code{prox_matrix} should be the name of the matrix component as a string.}

\item{proximity_type}{Character string specifying the type of proximity measure in \code{prox_matrix}. Options are \code{"similarity"} (higher values indicate closer proximity) or \code{"dissimilarity"} (lower values indicate closer proximity). Defaults to \code{"similarity"}.}

\item{method}{Character string specifying the silhouette calculation method. Options are \code{"pac"} (Probability of Alternative Cluster) or \code{"medoid"}. Defaults to \code{"medoid"}.}

\item{prob_matrix}{A numeric matrix where rows represent observations and columns represent cluster membership probabilities (or transformed probabilities, depending on \code{prob_type}). If \code{clust_fun} is provided, \code{prob_matrix} should be the name of the matrix component as a string (e.g., \code{"z"} for \code{Mclust}). When its not \code{NUll} fuzzy silhouette width is calculated. Defaults to \code{NULL} calculates crisp silhouette.}

\item{a}{Numeric value controlling the weight calculation based on membership difference. Must be positive. Defaults to \code{2}.}

\item{clust_fun}{Optional S3 or S4 function object specifying a clustering function that produces the proximity measure matrix. For example, \code{fanny}. If provided, \code{prox_matrix} must be the name of the matrix component in the clustering output (e.g., \code{"membership"} for \code{fanny} when  \code{proximity_type = "similarity"} ). Defaults to \code{NULL}.}

\item{...}{Additional arguments passed to \code{clust_fun}, such as \code{data,G} for \code{Mclust}.}
}
\value{
A list of class \code{"silhouette"} with the following components:
\describe{
\item{widths}{A data frame containing cluster assignments, nearest neighbor clusters, and silhouette widths for each observation.}
\item{clus.avg.widths}{A named numeric vector of average silhouette widths per cluster.}
\item{avg.width}{A numeric value representing the overall average silhouette width across all observations.}
}
}
\description{
Computes the silhouette width for each observation based on clustering results, measuring how similar an observation is to its own cluster compared to other clusters. The silhouette width ranges from -1 to 1, where higher values indicate better cluster cohesion and separation.
}
\examples{
# Example with iris dataset and k-means clustering
# kmeans clustering dont have disimilarity measure in its output
data(iris)
kmeans_result <- kmeans(iris[, 1:4], centers = 3)

# Compute distance matrix (observations x clusters)
distance_matrix <- as.matrix(dist(rbind(kmeans_result$centers, iris[, 1:4])))[-(1:3), 1:3]
out <- silhouette(prox_matrix = distance_matrix, proximity_type = "dissimilarity")

# Plot silhouette results
plot(out, label = FALSE, print.summary = TRUE)

# Example with clustering function
library(mclust)
out1 <- silhouette(prox_matrix = "z", clust_fun = "Mclust", data = iris[, 1:4], G = 3)
# Plot silhouette results
plot(out1, label = FALSE, print.summary = TRUE)


}
\references{
Rousseeuw, P. J. (1987). Silhouettes: A graphical aid to the interpretation and validation of cluster analysis. \emph{Journal of Computational and Applied Mathematics}, 20, 53--65. \href{https://doi.org/10.1016/0377-0427(87)90125-7}{DOI:10.1016/0377-0427(87)90125-7}

Van der Laan, M., Pollard, K., & Bryan, J. (2003). A new partitioning around medoids algorithm. \emph{Journal of Statistical Computation and Simulation}, 73(8), 575--584. \href{https://doi.org/10.1080/0094965031000136012}{DOI:10.1080/0094965031000136012}

Raymaekers, J., & Rousseeuw, P. J. (2022). Silhouettes and quasi residual plots for neural nets and tree-based classifiers. \emph{Journal of Computational and Graphical Statistics}, 31(4), 1332--1343. \href{https://doi.org/10.1080/10618600.2022.2050249}{DOI:10.1080/10618600.2022.2050249}

Campello, R. J., & Hruschka, E. R. (2006). A fuzzy extension of the silhouette width criterion for cluster analysis. \emph{Fuzzy Sets and Systems}, 157(21), 2858--2875. \href{https://doi.org/10.1016/j.fss.2006.07.006}{DOI:10.1016/j.fss.2006.07.006}

Schepers, J., Ceulemans, E., & Van Mechelen, I. (2008). Selecting among multi-mode partitioning models of different complexities: A comparison of four model selection criteria. \emph{Journal of Classification}, 25(1), 67--85. \href{https://doi.org/10.1007/s00357-008-9005-9}{DOI:10.1007/s00357-008-9005-9}
}
